refactor: import i18n path helpers directly from path module
